+++ packages/coin-support-canton/src/operations/createAccounts/index.ts
@@ -24,7 +24,7 @@ import {
 import * as services from '../../services';
 import { createApp } from '../../utils';
 
-const DERIVATION_PATH_LIMIT = 1;
+const DERIVATION_PATH_LIMIT = 20;
 
 const getAddressesFromDevice: GetAddressesFromDevice<
   CantonApp
@@ -106,22 +106,21 @@ const getBalanceAndTxnCount = async (
   publicKey: string,
   params: ICreateCantonAccountParams,
 ) => {
+  const coin = cantonCoinList[params.coinId];
   const partyId = await derivePartyId(hexToUint8Array(publicKey));
-  const isAccountCreated = await services.getIsAccountCreated(
-    partyId,
-    params.keyDB,
-  );
-
-  if (isAccountCreated) {
-    return {
-      balance: await services.getBalance(partyId, params.keyDB),
-      txnCount: 1,
-    };
-  }
 
   return {
-    balance: '0',
-    txnCount: 0,
+    balance: await services.getBalance(partyId, coin.instrument, params.keyDB),
+    txnCount: (
+      await services.getTransactions(
+        {
+          partyId,
+          fetchAll: true,
+          afterOffset: 0,
+        },
+        params.keyDB,
+      )
+    ).count,
   };
 };