+++ packages/coin-support-canton/src/operations/getExplorerLink/index.ts
@@ -5,7 +5,7 @@ import { config } from '../../config';
 
 export const getExplorerLink = (params: IGetExplorerLink) => {
   const queryParams = {
-    network: cantonCoinList[params.transaction.assetId].network,
+    network: cantonCoinList[params.transaction.parentAssetId].network,
     txHash: params.transaction.hash,
   };