+++ packages/coins/src/canton/coins.ts
@@ -7,6 +7,7 @@ export default [
     coinGeckoId: 'canton-network',
     coinIndex: '80001a6f', // 6767'
     feesUnit: 'CC',
+    decimals: 10,
     color: '#F3FE97',
     family: 'canton',
     network: 'mainnet',
@@ -17,5 +18,10 @@ export default [
         magnitude: 0,
       },
     ],
+    instrument: {
+      id: 'Amulet',
+      admin:
+        'DSO::1220b1431ef217342db44d516bb9befde802be7d8899637d290895fa58880f19accc',
+    },
   },
 ];