+++ packages/coins/src/canton/index.ts
@@ -1,12 +1,22 @@
 import coinList from './coins';
+import { getCantonTokens, ICantonToken } from './token';
 
 import { ICoinInfo, coinFamiliesMap } from '../types';
+
+export * from './token';
 
 type CantonFamily = typeof coinFamiliesMap.canton;
 
 export interface ICantonCoinInfo extends ICoinInfo {
   family: CantonFamily;
   network: string;
+  instrument: {
+    id: string;
+    admin: string;
+  };
+  decimals: number;
+  tokens: Record<string, ICantonToken>;
+  tokensByContract: Record<string, ICantonToken>;
 }
 
 export const CantonIdMap = {
@@ -29,9 +39,12 @@ export const cantonCoinList: Record<string, ICantonCoinInfo> = coinList.reduce<
       coinGeckoId: coin.coinGeckoId,
       coinIndex: coin.coinIndex,
       feesUnit: coin.feesUnit,
+      decimals: coin.decimals,
       network: coin.network,
       units: coin.units,
       color: coin.color,
+      instrument: coin.instrument,
+      ...getCantonTokens(coin.id, { color: coin.color }),
     },
   }),
   {},