+++ packages/cysync-core/src/dialogs/AddAccount/Dialogs/Congrats.tsx
@@ -1,4 +1,3 @@
-import { coinFamiliesMap } from '@cypherock/coins';
 import { ConfettiBlast, SuccessDialog } from '@cypherock/cysync-ui';
 import React from 'react';
 
@@ -10,15 +9,12 @@ export const AddAccountCongrats: React.FC = () => {
   const lang = useAppSelector(selectLanguage);
   const strings = lang.strings.addAccount.congrats;
   const button = lang.strings.buttons;
-  const { goTo, onClose, selectedCoin, setSelectedCoin } =
-    useAddAccountDialog();
+  const { goTo, onClose, setSelectedCoin } = useAddAccountDialog();
 
   const handleAddMoreAccount = () => {
     setSelectedCoin(undefined);
     goTo(0, 0);
   };
-
-  const isCanton = selectedCoin?.family === coinFamiliesMap.canton;
 
   return (
     <>
@@ -27,9 +23,9 @@ export const AddAccountCongrats: React.FC = () => {
         title={strings.title}
         subtext={strings.subtext}
         buttonText={button.done}
-        secondaryButtonText={!isCanton ? strings.buttonAddMore : undefined}
+        secondaryButtonText={strings.buttonAddMore}
         handleClick={onClose}
-        handleSecButtonClick={!isCanton ? handleAddMoreAccount : undefined}
+        handleSecButtonClick={handleAddMoreAccount}
       />
     </>
   );