+++ packages/core-services/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @cypherock/cysync-core-services
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [[`01b6f2e0`](https://github.com/Cypherock/cypherock-cysync/commit/01b6f2e08312190d23546ac7600e3a5b526c64c4), [`cc17bed6`](https://github.com/Cypherock/cypherock-cysync/commit/cc17bed6452b60393efe3fceb3e68868b3c088b2)]:
+  - @cypherock/coins@0.6.0
+  - @cypherock/coin-support-utils@0.1.5
+  - @cypherock/coin-support@0.2.7
+  - @cypherock/coin-support-evm@0.1.5
+
+## 0.2.7-pre.0
+
+### Patch Changes
+
+- Updated dependencies [[`01b6f2e0`](https://github.com/Cypherock/cypherock-cysync/commit/01b6f2e08312190d23546ac7600e3a5b526c64c4), [`cc17bed6`](https://github.com/Cypherock/cypherock-cysync/commit/cc17bed6452b60393efe3fceb3e68868b3c088b2)]:
+  - @cypherock/coins@0.6.0-pre.0
+  - @cypherock/coin-support-utils@0.1.5-pre.0
+  - @cypherock/coin-support@0.2.7-pre.0
+  - @cypherock/coin-support-evm@0.1.5-pre.0
+
 ## 0.2.6
 
 ### Patch Changes